Moved file handling / qr-code logic from main.py to file_handling.py
